requirements.txt
django-phonenumber-field==5.0.0
djangorestframework==3.10.3
Django==2.2.5